+++ Sources/CPIOArchiveKit/FileType.swift
@@ -1,13 +1,14 @@
-// Copyright (c) 2021 Jeff Lebrun
+// Copyright (c) 2024 Jeff Lebrun
 //
 //  Licensed under the MIT License.
 //
 //  The full text of the license can be found in the file named LICENSE.
 
-/// `FileType` specifies the type of file in an archive.
+/// A `CPIOFileType` specifies the type of file in an archive.
 ///
-/// Documentation comments are taken from [FreeBSD's man pages](https://www.freebsd.org/cgi/man.cgi?query=cpio&sektion=5&manpath=FreeBSD+13.0-current).
-public enum FileType: UInt32, Codable, Equatable, CaseIterable {
+/// Documentation comments are taken from [FreeBSD's man
+/// pages](https://www.freebsd.org/cgi/man.cgi?query=cpio&sektion=5&manpath=FreeBSD+13.0-current).
+public enum CPIOFileType: UInt32, Codable, Equatable, CaseIterable {
 	/// SUID bit.
 	case setUID = 0o4000
 
@@ -41,8 +42,8 @@ public enum FileType: UInt32, Codable, Equatable, CaseIterable {
 	/// This masks the file type bits.
 	///
 	/// ```swift
-	/// if (reader.headers[0].mode.rawValue & FileType.type.rawValue) == FileType.regular.rawValue {
-	/// 	// reader.headers[0] describes a file.
+	/// if (archive.files[0].header.mode.rawValue & CPIOFileType.type.rawValue) == CPIOFileType.regular.rawValue {
+	/// 	// reader.files[0] describes a file.
 	/// }
 	/// ```
 	case type = 0o170000
@@ -52,25 +53,26 @@ public enum FileType: UInt32, Codable, Equatable, CaseIterable {
 	case permissions = 0o777
 }
 
-/// A `FileMode` contains a file's permissions and file type.
-public struct FileMode: RawRepresentable, Equatable {
+/// A `CPIOFileMode` contains a file's permissions and file type.
+public struct CPIOFileMode: RawRepresentable, Equatable {
 	public typealias RawValue = UInt32
 
 	public var rawValue: UInt32
 
 	public init(rawValue: UInt32) { self.rawValue = rawValue }
 
-	/// Creates a `FileMode`.
+	/// Creates a `CPIOFileMode`.
 	/// - Parameters:
 	///   - permissions: The UNIX file permissions.
-	///   - modes: `FileType`s that describe the file.
-	/// 	For example, if you wanted to specify that this `FileMode` describes a directory, you would add the `.dir` type. If it is also a symlink, you would add the `.symlink` type.
-	public init(_ permissions: UInt32, modes: Set<FileType> = [.regular]) {
+	///   - modes: ``CPIOFileType``s that describe the file.
+	/// 	For example, if you wanted to specify that this `CPIOFileMode` describes a directory, you would add the
+	/// ``CPIOFileType/directory`` type. If it is also a symlink, you would add the ``CPIOFileType/symlink``  type.
+	public init(_ permissions: UInt32, modes: Set<CPIOFileType> = [.regular]) {
 		self.rawValue = permissions
 		modes.forEach({ self.rawValue |= $0.rawValue })
 	}
 
-	/// Check whether this `FileMode` is `type`.
+	/// Check whether this `CPIOFileMode` is `type`.
 	///
 	/// ```swift
 	/// if myHeader.mode.is(.directory) {
@@ -79,16 +81,16 @@ public struct FileMode: RawRepresentable, Equatable {
 	///     // myHeader describes a file.
 	/// } else if ...
 	/// ```
-	public func `is`(_ type: FileType) -> Bool {
-		(self.withoutPermissions & FileType.type.rawValue) == type.rawValue
+	public func `is`(_ type: CPIOFileType) -> Bool {
+		(self.withoutPermissions & CPIOFileType.type.rawValue) == type.rawValue
 	}
 
-	/// The UNIX permissions of this `FileMode`.
+	/// The UNIX permissions of this `CPIOFileMode`.
 	public var permissions: UInt32 {
-		self.rawValue & FileType.permissions.rawValue
+		self.rawValue & CPIOFileType.permissions.rawValue
 	}
 
-	/// The `FileType` of `self`.
+	/// The ``CPIOFileType`` of `self`.
 	///
 	/// ```swift
 	/// switch header.mode.type {
@@ -100,17 +102,17 @@ public struct FileMode: RawRepresentable, Equatable {
 	/// 		// unknown type.
 	/// }
 	/// ```
-	public var type: FileType? {
-		FileType(rawValue: self.rawType)
+	public var type: CPIOFileType? {
+		CPIOFileType(rawValue: self.rawType)
 	}
 
-	/// The file type bits in `self.rawValue`, or, in other words, the raw version of `self.type`.
+	/// The file type bits in ``CPIOFileMode/rawValue``, or, in other words, the raw version of ``CPIOFileMode/type``.
 	public var rawType: UInt32 {
-		self.withoutPermissions & FileType.type.rawValue
+		self.withoutPermissions & CPIOFileType.type.rawValue
 	}
 
-	/// `self.rawValue` without the UNIX permissions.
+	/// ``CPIOFileMode/rawValue`` without the UNIX permissions.
 	private var withoutPermissions: UInt32 {
-		self.rawValue &^ FileType.permissions.rawValue
+		self.rawValue &^ CPIOFileType.permissions.rawValue
 	}
 }