+++ Tests/CPIOArchiveKitTests/CPIOArchiveKitTests.swift
@@ -1,4 +1,4 @@
-// Copyright (c) 2021 Jeff Lebrun
+// Copyright (c) 2024 Jeff Lebrun
 //
 //  Licensed under the MIT License.
 //
@@ -9,175 +9,141 @@ import XCTest
 
 final class CPIOArchiveKitTests: XCTestCase {
 	func testWriteArchive() throws {
-		var writer = CPIOArchiveWriter()
-
-		writer.addFile(
-			header: Header(
-				name: "hello.txt",
-				mode: FileMode(0o644),
-				modificationTime: 1620311816,
-				links: 1
+		let archive = CPIOArchive([
+			.init(
+				header:
+				CPIOArchive.Header(
+					name: "hello.txt",
+					mode: CPIOFileMode(0o644),
+					modificationTime: 1620311816,
+					links: 1
+				),
+				contents: "Hello, World!\n"
 			),
-			contents: "Hello, World!\n"
-		)
-
-		writer.addFile(
-			header: Header(
-				name: "hello2.txt",
-				mode: FileMode(0o655),
-				modificationTime: 1620311816,
-				links: 1
+			.init(
+				header: CPIOArchive.Header(
+					name: "hello2.txt",
+					mode: CPIOFileMode(0o655),
+					modificationTime: 1620311816,
+					links: 1
+				),
+				contents: "Hello, Again!\n"
 			),
-			contents: "Hello, Again!\n"
-		)
 
-		writer.addFile(
-			header: Header(
-				name: "symlink.txt",
-				mode: FileMode(0o644, modes: [.symlink]),
-				modificationTime: 1620311816,
-				links: 1
+			.init(
+				header: CPIOArchive.Header(
+					name: "symlink.txt",
+					mode: CPIOFileMode(0o644, modes: [.symlink]),
+					modificationTime: 1620311816,
+					links: 1
+				),
+				contents: "hello.txt"
 			),
-			contents: "hello.txt"
-		)
 
-		writer.addFile(
-			header: Header(
-				name: "directory/",
-				mode: FileMode(0o644, modes: [.directory]),
-				modificationTime: 1620311816,
-				links: 1
-			)
-		)
+			.init(
+				header: CPIOArchive.Header(
+					name: "directory/",
+					mode: CPIOFileMode(0o644, modes: [.directory]),
+					modificationTime: 1620311816,
+					links: 1
+				)
+			),
+		])
 
-		XCTAssertEqual(Data(writer.finalize()), try Data(contentsOf: Bundle.module.url(forResource: "test-files/archive", withExtension: "cpio")!))
+		XCTAssertEqual(
+			Data(archive.serialize()),
+			try Data(contentsOf: Bundle.module.url(forResource: "test-files/archive", withExtension: "cpio")!)
+		)
 	}
 
 	func testReadArchive() throws {
 		let headers = [
-			Header(
+			CPIOArchive.Header(
 				name: "hello.txt",
-				mode: FileMode(0o644, modes: [.regular]),
+				mode: CPIOFileMode(0o644, modes: [.regular]),
 				modificationTime: 1620311816,
 				inode: 0,
 				links: 1,
 				dev: (major: 0, minor: 0),
 				checksum: Checksum(sum: 0x473)
 			),
-			Header(
+			CPIOArchive.Header(
 				name: "hello2.txt",
-				mode: FileMode(0o655, modes: [.regular]),
+				mode: CPIOFileMode(0o655, modes: [.regular]),
 				modificationTime: 1620311816,
 				inode: 1,
 				links: 1,
 				dev: (major: 1, minor: 1),
 				checksum: Checksum(sum: 0x44B)
 			),
-			Header(
+			CPIOArchive.Header(
 				name: "symlink.txt",
-				mode: FileMode(0o644, modes: [.symlink]),
+				mode: CPIOFileMode(0o644, modes: [.symlink]),
 				modificationTime: 1620311816,
 				inode: 2,
 				links: 1,
 				dev: (major: 2, minor: 2)
 			),
-			Header(
+			CPIOArchive.Header(
 				name: "directory/",
-				mode: FileMode(0o644, modes: [.directory]),
+				mode: CPIOFileMode(0o644, modes: [.directory]),
 				modificationTime: 1620311816,
 				inode: 3,
 				links: 1,
 				dev: (major: 3, minor: 3)
 			),
 		]
-		let reader = try CPIOArchiveReader(archive: Array(Data(contentsOf: Bundle.module.url(forResource: "test-files/archive", withExtension: "cpio")!)))
+		let archive = try CPIOArchive(data: Array(Data(contentsOf: Bundle.module.url(
+			forResource: "test-files/archive",
+			withExtension: "cpio"
+		)!)))
 
-		XCTAssertEqual(reader.headers[0].name, headers[0].name)
-		XCTAssertEqual(reader.headers[0].mode, headers[0].mode)
-		XCTAssertEqual(reader.headers[0].userID, headers[0].userID)
-		XCTAssertEqual(reader.headers[0].groupID, headers[0].groupID)
-		XCTAssertEqual(reader.headers[0].modificationTime, headers[0].modificationTime)
-		XCTAssertEqual(reader.headers[0].inode, headers[0].inode)
-		XCTAssertEqual(reader.headers[0].dev.major, headers[0].dev.major)
-		XCTAssertEqual(reader.headers[0].dev.minor, headers[0].dev.minor)
-		XCTAssertEqual(reader.headers[0].rDev.major, headers[0].rDev.major)
-		XCTAssertEqual(reader.headers[0].rDev.minor, headers[0].rDev.minor)
-		XCTAssertEqual(headers[0].checksum!.sum, Checksum(bytes: reader[0]).sum)
+		XCTAssertEqual(archive.files.count, headers.count)
 
-		XCTAssertEqual(reader.headers[1].name, headers[1].name)
-		XCTAssertEqual(reader.headers[1].mode, headers[1].mode)
-		XCTAssertEqual(reader.headers[1].userID, headers[1].userID)
-		XCTAssertEqual(reader.headers[1].groupID, headers[1].groupID)
-		XCTAssertEqual(reader.headers[1].modificationTime, headers[1].modificationTime)
-		XCTAssertEqual(reader.headers[1].inode, headers[1].inode)
-		XCTAssertEqual(reader.headers[1].dev.major, headers[1].dev.major)
-		XCTAssertEqual(reader.headers[1].dev.minor, headers[1].dev.minor)
-		XCTAssertEqual(reader.headers[1].rDev.major, headers[1].rDev.major)
-		XCTAssertEqual(reader.headers[1].rDev.minor, headers[1].rDev.minor)
-		XCTAssertEqual(headers[1].checksum!.sum, Checksum(bytes: reader[1]).sum)
-
-		XCTAssertEqual(reader.headers[2].name, headers[2].name)
-		XCTAssertEqual(reader.headers[2].mode, headers[2].mode)
-		XCTAssertEqual(reader.headers[2].userID, headers[2].userID)
-		XCTAssertEqual(reader.headers[2].groupID, headers[2].groupID)
-		XCTAssertEqual(reader.headers[2].modificationTime, headers[2].modificationTime)
-		XCTAssertEqual(reader.headers[2].inode, headers[2].inode)
-		XCTAssertEqual(reader.headers[2].dev.major, headers[2].dev.major)
-		XCTAssertEqual(reader.headers[2].dev.minor, headers[2].dev.minor)
-		XCTAssertEqual(reader.headers[2].rDev.major, headers[2].rDev.major)
-		XCTAssertEqual(reader.headers[2].rDev.minor, headers[2].rDev.minor)
-
-		XCTAssertEqual(reader.headers[3].name, headers[3].name)
-		XCTAssertEqual(reader.headers[3].mode, headers[3].mode)
-		XCTAssertEqual(reader.headers[3].userID, headers[3].userID)
-		XCTAssertEqual(reader.headers[3].groupID, headers[3].groupID)
-		XCTAssertEqual(reader.headers[3].modificationTime, headers[3].modificationTime)
-		XCTAssertEqual(reader.headers[3].inode, headers[3].inode)
-		XCTAssertEqual(reader.headers[3].dev.major, headers[3].dev.major)
-		XCTAssertEqual(reader.headers[3].dev.minor, headers[3].dev.minor)
-		XCTAssertEqual(reader.headers[3].rDev.major, headers[3].rDev.major)
-		XCTAssertEqual(reader.headers[3].rDev.minor, headers[3].rDev.minor)
+		for i in 0..<headers.count {
+			XCTAssertEqual(archive.files[i].header.name, headers[i].name)
+			XCTAssertEqual(archive.files[i].header.mode, headers[i].mode)
+			XCTAssertEqual(archive.files[i].header.userID, headers[i].userID)
+			XCTAssertEqual(archive.files[i].header.groupID, headers[i].groupID)
+			XCTAssertEqual(archive.files[i].header.modificationTime, headers[i].modificationTime)
+			XCTAssertEqual(archive.files[i].header.inode, headers[i].inode)
+			XCTAssertEqual(archive.files[i].header.dev.major, headers[i].dev.major)
+			XCTAssertEqual(archive.files[i].header.dev.minor, headers[i].dev.minor)
+			XCTAssertEqual(archive.files[i].header.rDev.major, headers[i].rDev.major)
+			XCTAssertEqual(archive.files[i].header.rDev.minor, headers[i].rDev.minor)
+			if let checksum = headers[i].checksum {
+				XCTAssertEqual(Checksum(bytes: archive.files[i].contents).sum, checksum.sum)
+			}
+		}
 	}
 
 	func testReadBIGArchive() throws {
-		let header = Header(
+		let header = CPIOArchive.Header(
 			name: "large.txt",
 			userID: 0,
 			groupID: 0,
-			mode: FileMode(0o644, modes: [.regular]),
+			mode: CPIOFileMode(0o644, modes: [.regular]),
 			modificationTime: 1621215500,
 			inode: 32430807,
 			links: 1,
 			dev: (major: 0, minor: 121)
 		)
 
-		let reader = try CPIOArchiveReader(archive: Array(Data(contentsOf: Bundle.module.url(forResource: "test-files/big", withExtension: "cpio")!)))
+		let archive = try CPIOArchive(data: Array(Data(contentsOf: Bundle.module.url(
+			forResource: "test-files/big",
+			withExtension: "cpio"
+		)!)))
 
-		XCTAssertEqual(reader.headers[0].name, header.name)
-		XCTAssertEqual(reader.headers[0].mode, header.mode)
-		XCTAssertEqual(reader.headers[0].userID, header.userID)
-		XCTAssertEqual(reader.headers[0].groupID, header.groupID)
-		XCTAssertEqual(reader.headers[0].mode, header.mode)
-		XCTAssertEqual(reader.headers[0].modificationTime, header.modificationTime)
-		XCTAssertEqual(reader.headers[0].inode, header.inode)
-		XCTAssertEqual(reader.headers[0].dev.major, header.dev.major)
-		XCTAssertEqual(reader.headers[0].dev.minor, header.dev.minor)
-		XCTAssertEqual(reader.headers[0].rDev.major, header.rDev.major)
-		XCTAssertEqual(reader.headers[0].rDev.minor, header.rDev.minor)
+		XCTAssertEqual(archive.files[0].header.name, header.name)
+		XCTAssertEqual(archive.files[0].header.mode, header.mode)
+		XCTAssertEqual(archive.files[0].header.userID, header.userID)
+		XCTAssertEqual(archive.files[0].header.groupID, header.groupID)
+		XCTAssertEqual(archive.files[0].header.mode, header.mode)
+		XCTAssertEqual(archive.files[0].header.modificationTime, header.modificationTime)
+		XCTAssertEqual(archive.files[0].header.inode, header.inode)
+		XCTAssertEqual(archive.files[0].header.dev.major, header.dev.major)
+		XCTAssertEqual(archive.files[0].header.dev.minor, header.dev.minor)
+		XCTAssertEqual(archive.files[0].header.rDev.major, header.rDev.major)
+		XCTAssertEqual(archive.files[0].header.rDev.minor, header.rDev.minor)
 	}
-
-	func testReaderIterator() throws {
-		let reader = try CPIOArchiveReader(archive: Array(Data(contentsOf: Bundle.module.url(forResource: "test-files/archive", withExtension: "cpio")!)))
-
-		for (_, _) in reader {
-			// This shouldn't crash
-		}
-	}
-
-	static var allTests = [
-		("Test Writing An Archive", testWriteArchive),
-		("Test Reading An Archive", testReadArchive),
-		("Test Reading A BIG Archive", testReadBIGArchive),
-		("Test Iterator", testReaderIterator),
-	]
 }